requirements.txt
alembic==1.6.5
APScheduler==3.7.0
cffi==1.14.5
click==8.0.1
cryptography==3.4.7
Flask==2.0.1
Flask-Migrate==3.0.1
flask-sqlacodegen==1.1.8
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
inflect==5.3.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
pycparser==2.20
PyMySQL==1.0.2
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.17
tzlocal==2.1
Werkzeug==2.0.1
